Reject trials acquired during setup issue
Issue during acquisition (e.g. port malfunction) is manually reported by the experimenter on a Google Sheet. This info is then loaded by the routine and the trials acquired during the respective periods are automatically rejected and not pushed to the DataJoint table.

Also, we add new fields to the table corresponding to EEG recordings and the new beta parameter for v3 protocol.